+++ sample/src/main/java/com/origeek/viewerDemo/GalleryActivity.kt
@@ -6,7 +6,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.mutableStateListOf
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
-import com.origeek.imageViewer.ImageGallery
+import com.origeek.imageViewer.gallery.ImageGallery
 import com.origeek.viewerDemo.base.BaseActivity
 import com.origeek.viewerDemo.ui.component.rememberCoilImagePainter
 import com.origeek.viewerDemo.ui.theme.ViewerDemoTheme
